Style for 'List of Talks' page
Change the table layout for a responsive layout that looks better in
mobile devices. Also, removing some texts and options.

Talk admins are no longer able to confirm/unconfirm talks from this
page, wich I think is a positive thing because the two checkboxes (add
to myschedule/confirm) were confusing and was not clear what the submit
button did.

I'm also removing the 'my schedule' checkboxes. I've had a hard time
trying to make that work with the javascript provided by ACT. Because
it's a small conference I think we can live without it. We can try to
add it later if we feel brave enough.